chore(): put FREQUENCY_OPTIONS to /constants/beep
