+++ src/constants/beep.ts
@@ -1,0 +1,6 @@
+export const FREQUENCY_OPTIONS = [
+  { frequency: '1', name: 'Daily' },
+  { frequency: '7', name: 'Weekly' },
+  { frequency: '14', name: 'Bi-Weekly' },
+  { frequency: '30', name: 'Monthly' },
+]
